testes-trabalho/pyproject.toml
[tool.poetry]
name = "testes_trabalho"
version = "0.1.0"
description = ""
authors = ["Luís Simas <luissimas@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pyzmq = "^26.0.2"
opencv-python = "^4.9.0.80"

[tool.pyright]
exclude=[ ".venv" ]
venvPath="."
venv=".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
